5047 Remove splitterv2 module
