+++ .github/config/chunks.yaml
@@ -52,7 +52,6 @@ chunks:
       - compaction/compaction-tracker
       - compaction/compaction-core
       - partitions/splitter
-      - partitions/splitterv2
       - partitions/splitter-lambda
       - garbage-collector
       - bulk-export/bulk-export-core
