+++ .github/workflows/chunk-clients-cdk.yaml
@@ -32,7 +32,7 @@ on:
       - 'java/configurationV2/**'
       - 'java/common/dynamodb-toolsv2/**'
       - 'java/bulk-import/bulk-import-core/**'
-      - 'java/partitions/splitterv2/**'
+      - 'java/partitions/splitter/**'
       - 'java/compaction/compaction-job-execution/**'
       - 'java/statestore-committer-core/**'
       - 'java/garbage-collector/**'
