+++ .github/workflows/chunk-compaction.yaml
@@ -18,7 +18,6 @@ on:
       - 'java/compaction/compaction-tracker/**'
       - 'java/compaction/compaction-core/**'
       - 'java/partitions/splitter/**'
-      - 'java/partitions/splitterv2/**'
       - 'java/partitions/splitter-lambda/**'
       - 'java/garbage-collector/**'
       - 'java/bulk-export/bulk-export-core/**'
