+++ .github/workflows/chunk-query.yaml
@@ -16,7 +16,7 @@ on:
       - 'java/query/query-lambdav2/**'
       - 'java/athena/**'
       - 'java/trino/**'
-      - 'java/partitions/splitterv2/**'
+      - 'java/partitions/splitter/**'
       - 'java/ingest/ingest-runner/**'
       - 'java/statestore/**'
       - 'java/statestorev2/**'
