+++ code-style/checkstyle-suppressions.xml
@@ -72,8 +72,6 @@
     <suppress files="[\\/]sleeper[\\/]compaction[\\/]tracker[\\/].*" id="missingJavadocType"/>
     <suppress files="[\\/]sleeper[\\/]splitter[\\/].*" id="missingJavadocMethod"/>
     <suppress files="[\\/]sleeper[\\/]splitter[\\/].*" id="missingJavadocType"/>
-    <suppress files="[\\/]sleeper[\\/]splitterv2[\\/].*" id="missingJavadocMethod"/>
-    <suppress files="[\\/]sleeper[\\/]splitterv2[\\/].*" id="missingJavadocType"/>
     <suppress files="[\\/]sleeper[\\/]garbagecollector[\\/].*" id="missingJavadocMethod"/>
     <suppress files="[\\/]sleeper[\\/]garbagecollector[\\/].*" id="missingJavadocType"/>
     <suppress files="[\\/]sleeper[\\/]ingest[\\/]batcher[\\/].*" id="missingJavadocMethod"/>
