+++ java/athena/pom.xml
@@ -78,7 +78,7 @@
         </dependency>
         <dependency>
             <groupId>sleeper</groupId>
-            <artifactId>splitterv2</artifactId>
+            <artifactId>splitter</artifactId>
             <version>${project.parent.version}</version>
             <scope>test</scope>
         </dependency>
