+++ java/athena/src/test/java/sleeper/athena/metadata/SleeperMetadataHandlerIT.java
@@ -52,7 +52,7 @@ import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.core.properties.table.TableProperties;
 import sleeper.core.statestore.StateStore;
 import sleeper.sketchesv2.store.LocalFileSystemSketchesStore;
-import sleeper.splitterv2.core.split.SplitPartition;
+import sleeper.splitter.core.split.SplitPartition;
 import sleeper.statestorev2.StateStoreFactory;
 
 import java.io.IOException;
