+++ java/clients/src/main/java/sleeper/clients/report/DeadLettersStatusReport.java
@@ -28,7 +28,7 @@ import sleeper.configurationv2.properties.S3TableProperties;
 import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.core.properties.table.TablePropertiesProvider;
 import sleeper.query.core.model.QuerySerDe;
-import sleeper.splitterv2.core.find.SplitPartitionJobDefinitionSerDe;
+import sleeper.splitter.core.find.SplitPartitionJobDefinitionSerDe;
 
 import java.util.function.Function;
