+++ java/clients/src/main/java/sleeper/clients/report/PartitionsStatusReport.java
@@ -22,7 +22,7 @@ import sleeper.clients.report.partitions.PartitionsStatusReportArguments;
 import sleeper.clients.report.partitions.PartitionsStatusReporter;
 import sleeper.core.properties.table.TableProperties;
 import sleeper.core.statestore.StateStore;
-import sleeper.splitterv2.core.status.PartitionsStatus;
+import sleeper.splitter.core.status.PartitionsStatus;
 
 import static sleeper.configurationv2.utils.AwsV2ClientHelper.buildAwsV2Client;
