+++ java/partitions/splitter-lambda/src/test/java/sleeper/splitter/lambda/FindPartitionsToSplitIT.java
@@ -43,9 +43,9 @@ import sleeper.ingest.runner.impl.partitionfilewriter.DirectPartitionFileWriterF
 import sleeper.ingest.runner.impl.recordbatch.arraylist.ArrayListRecordBatchFactory;
 import sleeper.localstack.test.LocalStackTestBase;
 import sleeper.sketchesv2.store.LocalFileSystemSketchesStore;
-import sleeper.splitterv2.core.find.FindPartitionsToSplit;
-import sleeper.splitterv2.core.find.SplitPartitionJobDefinition;
-import sleeper.splitterv2.core.find.SplitPartitionJobDefinitionSerDe;
+import sleeper.splitter.core.find.FindPartitionsToSplit;
+import sleeper.splitter.core.find.SplitPartitionJobDefinition;
+import sleeper.splitter.core.find.SplitPartitionJobDefinitionSerDe;
 
 import java.io.File;
 import java.io.IOException;
