--- java/partitions/splitterv2/src/main/java/sleeper/splitterv2/core/split/SplitPartitionResult.java
@@ -1,43 +1,0 @@
-/*
- * Copyright 2022-2025 Crown Copyright
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package sleeper.splitterv2.core.split;
-
-import sleeper.core.partition.Partition;
-
-public class SplitPartitionResult {
-
-    private final Partition parentPartition;
-    private final Partition leftChild;
-    private final Partition rightChild;
-
-    public SplitPartitionResult(Partition parentPartition, Partition leftChild, Partition rightChild) {
-        this.parentPartition = parentPartition;
-        this.leftChild = leftChild;
-        this.rightChild = rightChild;
-    }
-
-    public Partition getParentPartition() {
-        return parentPartition;
-    }
-
-    public Partition getLeftChild() {
-        return leftChild;
-    }
-
-    public Partition getRightChild() {
-        return rightChild;
-    }
-}
