--- java/partitions/splitterv2/src/test/java/sleeper/splitterv2/core/split/SplitPartitionTest.java
@@ -1,609 +1,0 @@
-/*
- * Copyright 2022-2025 Crown Copyright
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package sleeper.splitterv2.core.split;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Nested;
-import org.junit.jupiter.api.Test;
-
-import sleeper.core.partition.Partition;
-import sleeper.core.partition.PartitionTree;
-import sleeper.core.partition.PartitionsBuilder;
-import sleeper.core.properties.instance.InstanceProperties;
-import sleeper.core.properties.table.TableProperties;
-import sleeper.core.record.Record;
-import sleeper.core.schema.Field;
-import sleeper.core.schema.Schema;
-import sleeper.core.schema.type.ByteArrayType;
-import sleeper.core.schema.type.IntType;
-import sleeper.core.schema.type.LongType;
-import sleeper.core.schema.type.StringType;
-import sleeper.core.statestore.FileReference;
-import sleeper.core.statestore.FileReferenceFactory;
-import sleeper.core.statestore.StateStore;
-import sleeper.core.statestore.commit.StateStoreCommitRequest;
-import sleeper.core.statestore.testutils.InMemoryTransactionLogStateStore;
-import sleeper.core.statestore.testutils.InMemoryTransactionLogs;
-import sleeper.core.statestore.transactionlog.transaction.impl.SplitPartitionTransaction;
-import sleeper.sketchesv2.Sketches;
-import sleeper.sketchesv2.testutils.InMemorySketchesStore;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Map;
-import java.util.UUID;
-import java.util.concurrent.atomic.AtomicLong;
-import java.util.function.Supplier;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
-import java.util.stream.LongStream;
-import java.util.stream.Stream;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static sleeper.core.properties.table.TableProperty.PARTITION_SPLIT_ASYNC_COMMIT;
-import static sleeper.core.properties.table.TableProperty.TABLE_ID;
-import static sleeper.core.properties.testutils.InstancePropertiesTestHelper.createTestInstanceProperties;
-import static sleeper.core.properties.testutils.TablePropertiesTestHelper.createTestTableProperties;
-import static sleeper.core.schema.SchemaTestHelper.createSchemaWithKey;
-import static sleeper.core.statestore.testutils.StateStoreUpdatesWrapper.update;
-
-public class SplitPartitionTest {
-    private final InstanceProperties instanceProperties = createTestInstanceProperties();
-    private final TableProperties tableProperties = createTestTableProperties(instanceProperties, createSchemaWithKey("key", new IntType()));
-
-    private final List<StateStoreCommitRequest> sentAsyncCommits = new ArrayList<>();
-    private final InMemorySketchesStore filenameToSketches = new InMemorySketchesStore();
-
-    @Nested
-    @DisplayName("Skip split")
-    class SkipSplit {
-        @Test
-        public void shouldNotSplitPartitionForIntKeyIfItCannotBeSplitBecausePartitionIsOnePoint() throws Exception {
-            // Given
-            PartitionTree tree = new PartitionsBuilder(tableProperties)
-                    .rootFirst("root")
-                    .splitToNewChildren("root", "id12", "id3", 1)
-                    .splitToNewChildren("id12", "id1", "id2", 0)
-                    .buildTree();
-            StateStore stateStore = initialiseStateStore(tree.getAllPartitions());
-
-            for (Partition partition : tree.getAllPartitions()) {
-                int minRange = (int) partition.getRegion().getRange("key").getMin();
-                int maxRange = null == partition.getRegion().getRange("key").getMax() ? Integer.MAX_VALUE : (int) partition.getRegion().getRange("key").getMax();
-                for (int i = 0; i < 10; i++) {
-                    List<Record> records = new ArrayList<>();
-                    int j = 0;
-                    for (int r = minRange; r < maxRange && j < 10; r++, j++) {
-                        Record record = new Record();
-                        record.put("key", r);
-                        records.add(record);
-                    }
-                    ingestRecordsToSketchOnPartition(stateStore, partition.getId(), records.stream());
-                }
-            }
-
-            // When
-            splitPartition(stateStore, "id2", generateNoIds());
-
-            // Then
-            assertThat(stateStore.getAllPartitions())
-                    .containsExactlyInAnyOrderElementsOf(tree.getAllPartitions());
-        }
-
-        @Test
-        public void shouldNotSplitPartitionForIntKeyIfItCannotBeSplitBecauseDataIsConstant() throws Exception {
-            // Given
-            PartitionTree tree = new PartitionsBuilder(tableProperties)
-                    .rootFirst("root")
-                    .splitToNewChildren("root", "id12", "id3", 10)
-                    .splitToNewChildren("id12", "id1", "id2", 0)
-                    .buildTree();
-            StateStore stateStore = initialiseStateStore(tree.getAllPartitions());
-
-            for (Partition partition : stateStore.getAllPartitions()) {
-                for (int i = 0; i < 10; i++) {
-                    int minRange = (int) partition.getRegion().getRange("key").getMin();
-                    int maxRange = null == partition.getRegion().getRange("key").getMax() ? Integer.MAX_VALUE : (int) partition.getRegion().getRange("key").getMax();
-                    List<Record> records = new ArrayList<>();
-                    int j = 0;
-                    if (!partition.getId().equals("id2")) {
-                        for (int r = minRange; r < maxRange && j < 10; r++, j++) {
-                            Record record = new Record();
-                            record.put("key", r);
-                            records.add(record);
-                        }
-                    } else {
-                        // Files in partition2 all have the same value for the key
-                        for (int r = 0; r < 10; r++) {
-                            Record record = new Record();
-                            record.put("key", 1);
-                            records.add(record);
-                        }
-                    }
-                    ingestRecordsToSketchOnPartition(stateStore, partition.getId(), records.stream());
-                }
-            }
-
-            // When
-            splitPartition(stateStore, "id2", generateNoIds());
-
-            // Then
-            assertThat(stateStore.getAllPartitions())
-                    .containsExactlyInAnyOrderElementsOf(tree.getAllPartitions());
-        }
-
-        @Test
-        public void shouldNotSplitPartitionForByteArrayKeyIfItCannotBeSplitBecausePartitionIsOnePoint() throws Exception {
-            // Given
-            tableProperties.setSchema(Schema.builder().rowKeyFields(new Field("key", new ByteArrayType())).build());
-            PartitionTree tree = new PartitionsBuilder(tableProperties)
-                    .rootFirst("root")
-                    .splitToNewChildren("root", "id12", "id3", new byte[]{(byte) 51})
-                    .splitToNewChildren("id12", "id1", "id2", new byte[]{(byte) 50})
-                    .buildTree();
-
-            StateStore stateStore = initialiseStateStore(tree.getAllPartitions());
-
-            for (Partition partition : tree.getAllPartitions()) {
-                for (int i = 0; i < 10; i++) {
-                    List<Record> records = new ArrayList<>();
-                    if (partition.getId().equals("id1")) {
-                        int j = 0;
-                        for (byte r = (byte) 0;
-                             r < ((byte[]) partition.getRegion().getRange("key").getMax())[0] && j < 10;
-                             r++, j++) {
-                            Record record = new Record();
-                            record.put("key", new byte[]{r});
-                            records.add(record);
-                        }
-                    } else if (partition.getId().equals("id2")) {
-                        for (int j = 0; j < 10; j++) {
-                            Record record = new Record();
-                            record.put("key", new byte[]{50});
-                            records.add(record);
-                        }
-                    } else {
-                        for (int j = 51; j < 60; j++) {
-                            Record record = new Record();
-                            record.put("key", new byte[]{(byte) j});
-                            records.add(record);
-                        }
-                    }
-                    ingestRecordsToSketchOnPartition(stateStore, partition.getId(), records.stream());
-                }
-            }
-
-            // When
-            splitPartition(stateStore, "id2", generateNoIds());
-
-            // Then
-            assertThat(stateStore.getAllPartitions())
-                    .containsExactlyInAnyOrderElementsOf(tree.getAllPartitions());
-        }
-
-        @Test
-        public void shouldNotSplitPartitionForByteArrayKeyIfItCannotBeSplitBecauseDataIsConstant() throws Exception {
-            // Given
-            tableProperties.setSchema(Schema.builder().rowKeyFields(new Field("key", new ByteArrayType())).build());
-            PartitionTree tree = new PartitionsBuilder(tableProperties)
-                    .rootFirst("root")
-                    .splitToNewChildren("root", "id12", "id3", new byte[]{(byte) 100})
-                    .splitToNewChildren("id12", "id1", "id2", new byte[]{(byte) 50})
-                    .buildTree();
-            StateStore stateStore = initialiseStateStore(tree.getAllPartitions());
-
-            for (Partition partition : stateStore.getAllPartitions()) {
-                for (int i = 0; i < 10; i++) {
-                    List<Record> records = new ArrayList<>();
-                    if (partition.getId().equals("id1")) {
-                        int j = 0;
-                        for (byte r = (byte) 0;
-                             r < ((byte[]) partition.getRegion().getRange("key").getMax())[0] && j < 10;
-                             r++, j++) {
-                            Record record = new Record();
-                            record.put("key", new byte[]{r});
-                            records.add(record);
-                        }
-                    } else if (partition.getId().equals("id2")) {
-                        // Files in partition2 all have the same value for the key
-                        for (int j = 0; j < 10; j++) {
-                            Record record = new Record();
-                            record.put("key", new byte[]{60});
-                            records.add(record);
-                        }
-                    } else {
-                        for (int j = 100; j < 110; j++) {
-                            Record record = new Record();
-                            record.put("key", new byte[]{(byte) j});
-                            records.add(record);
-                        }
-                    }
-                    ingestRecordsToSketchOnPartition(stateStore, partition.getId(), records.stream());
-                }
-            }
-
-            // When
-            splitPartition(stateStore, "id2", generateNoIds());
-
-            // Then
-            assertThat(stateStore.getAllPartitions())
-                    .containsExactlyInAnyOrderElementsOf(tree.getAllPartitions());
-        }
-    }
-
-    @Nested
-    @DisplayName("Single dimension split")
-    class SingleDimensionSplit {
-        @Test
-        void shouldSplitPartitionForIntKey() throws Exception {
-            // Given
-            StateStore stateStore = initialiseStateStore(new PartitionsBuilder(tableProperties)
-                    .singlePartition("A")
-                    .buildList());
-            IntStream.range(0, 10)
-                    .forEach(i -> ingestRecordsToSketchOnPartition(stateStore, "A",
-                            IntStream.range(100 * i, 100 * (i + 1))
-                                    .mapToObj(r -> new Record(Map.of("key", r)))));
-
-            // When
-            splitSinglePartition(stateStore, generateIds("B", "C"));
-
-            // Then
-            assertThat(stateStore.getAllPartitions())
-                    .containsExactlyInAnyOrderElementsOf(new PartitionsBuilder(tableProperties)
-                            .rootFirst("A")
-                            .splitToNewChildren("A", "B", "C", 500)
-                            .buildList());
-        }
-
-        @Test
-        void shouldSplitPartitionForLongKey() throws Exception {
-            // Given
-            tableProperties.setSchema(Schema.builder().rowKeyFields(new Field("key", new LongType())).build());
-            StateStore stateStore = initialiseStateStore(new PartitionsBuilder(tableProperties)
-                    .singlePartition("A")
-                    .buildList());
-            IntStream.range(0, 10)
-                    .forEach(i -> ingestRecordsToSketchOnPartition(stateStore, "A",
-                            LongStream.range(100L * i, 100L * (i + 1))
-                                    .mapToObj(r -> new Record(Map.of("key", r)))));
-
-            // When
-            splitSinglePartition(stateStore, generateIds("B", "C"));
-
-            // Then
-            assertThat(stateStore.getAllPartitions())
-                    .containsExactlyInAnyOrderElementsOf(new PartitionsBuilder(tableProperties)
-                            .rootFirst("A")
-                            .splitToNewChildren("A", "B", "C", 500L)
-                            .buildList());
-        }
-
-        @Test
-        void shouldSplitPartitionForStringKey() throws Exception {
-            // Given
-            tableProperties.setSchema(Schema.builder().rowKeyFields(new Field("key", new StringType())).build());
-            StateStore stateStore = initialiseStateStore(new PartitionsBuilder(tableProperties)
-                    .singlePartition("A")
-                    .buildList());
-            IntStream.range(0, 10)
-                    .forEach(i -> ingestRecordsToSketchOnPartition(stateStore, "A",
-                            IntStream.range(0, 100)
-                                    .mapToObj(r -> new Record(Map.of("key", String.format("A%1d%02d", i, r))))));
-
-            // When
-            splitSinglePartition(stateStore, generateIds("B", "C"));
-
-            // Then
-            assertThat(stateStore.getAllPartitions())
-                    .containsExactlyInAnyOrderElementsOf(new PartitionsBuilder(tableProperties)
-                            .rootFirst("A")
-                            .splitToNewChildren("A", "B", "C", "A500")
-                            .buildList());
-        }
-
-        @Test
-        void shouldSplitPartitionForByteArrayKey() throws Exception {
-            // Given
-            tableProperties.setSchema(Schema.builder().rowKeyFields(new Field("key", new ByteArrayType())).build());
-            StateStore stateStore = initialiseStateStore(new PartitionsBuilder(tableProperties)
-                    .singlePartition("A")
-                    .buildList());
-            IntStream.range(0, 10)
-                    .forEach(i -> ingestRecordsToSketchOnPartition(stateStore, "A",
-                            IntStream.range(0, 100)
-                                    .mapToObj(r -> new Record(Map.of("key", new byte[]{(byte) r})))));
-
-            // When
-            splitSinglePartition(stateStore, generateIds("B", "C"));
-
-            // Then
-            assertThat(stateStore.getAllPartitions())
-                    .containsExactlyInAnyOrderElementsOf(new PartitionsBuilder(tableProperties)
-                            .rootFirst("A")
-                            .splitToNewChildren("A", "B", "C", new byte[]{(byte) 50})
-                            .buildList());
-        }
-    }
-
-    @Nested
-    @DisplayName("Multidimensional split")
-    class MultidimensionalSplit {
-        @Test
-        public void shouldSplitIntKeyOnFirstDimension() throws Exception {
-            // Given
-            tableProperties.setSchema(Schema.builder()
-                    .rowKeyFields(new Field("key1", new IntType()), new Field("key2", new IntType()))
-                    .build());
-            StateStore stateStore = initialiseStateStore(new PartitionsBuilder(tableProperties)
-                    .singlePartition("A")
-                    .buildList());
-            IntStream.range(0, 10)
-                    .forEach(i -> ingestRecordsToSketchOnPartition(stateStore, "A",
-                            IntStream.range(0, 100)
-                                    .mapToObj(r -> new Record(Map.of(
-                                            "key1", r,
-                                            "key2", 10)))));
-
-            // When
-            splitSinglePartition(stateStore, generateIds("B", "C"));
-
-            // Then
-            assertThat(stateStore.getAllPartitions())
-                    .containsExactlyInAnyOrderElementsOf(new PartitionsBuilder(tableProperties)
-                            .rootFirst("A")
-                            .splitToNewChildrenOnDimension("A", "B", "C", 0, 50)
-                            .buildList());
-        }
-
-        @Test
-        public void shouldSplitIntKeyOnSecondDimensionWhenAllValuesForFirstKeyAreTheSame() throws Exception {
-            // Given
-            tableProperties.setSchema(Schema.builder()
-                    .rowKeyFields(new Field("key1", new IntType()), new Field("key2", new IntType()))
-                    .build());
-            StateStore stateStore = initialiseStateStore(new PartitionsBuilder(tableProperties)
-                    .singlePartition("A")
-                    .buildList());
-            IntStream.range(0, 10)
-                    .forEach(i -> ingestRecordsToSketchOnPartition(stateStore, "A",
-                            IntStream.range(0, 100)
-                                    .mapToObj(r -> new Record(Map.of(
-                                            "key1", 10,
-                                            "key2", r)))));
-
-            // When
-            splitSinglePartition(stateStore, generateIds("B", "C"));
-
-            // Then
-            assertThat(stateStore.getAllPartitions())
-                    .containsExactlyInAnyOrderElementsOf(new PartitionsBuilder(tableProperties)
-                            .rootFirst("A")
-                            .splitToNewChildrenOnDimension("A", "B", "C", 1, 50)
-                            .buildList());
-        }
-
-        @Test
-        public void shouldSplitIntKeyOnFirstDimensionWhenSecondDimensionCanBeSplit() throws Exception {
-            // Given
-            tableProperties.setSchema(Schema.builder()
-                    .rowKeyFields(new Field("key1", new IntType()), new Field("key2", new IntType()))
-                    .build());
-            StateStore stateStore = initialiseStateStore(new PartitionsBuilder(tableProperties)
-                    .singlePartition("A")
-                    .buildList());
-            IntStream.range(0, 10)
-                    .forEach(i -> ingestRecordsToSketchOnPartition(stateStore, "A",
-                            IntStream.range(0, 100)
-                                    .mapToObj(r -> new Record(Map.of(
-                                            "key1", r,
-                                            "key2", i)))));
-
-            // When
-            splitSinglePartition(stateStore, generateIdsStartingFrom('B'));
-
-            // Then
-            assertThat(stateStore.getAllPartitions())
-                    .containsExactlyInAnyOrderElementsOf(new PartitionsBuilder(tableProperties)
-                            .rootFirst("A")
-                            .splitToNewChildrenOnDimension("A", "B", "C", 0, 50)
-                            .buildList());
-        }
-
-        @Test
-        public void shouldSplitIntKeyOnSecondDimensionWhenMinAndMedianForFirstKeyAreTheSame() throws Exception {
-            // Given
-            tableProperties.setSchema(Schema.builder()
-                    .rowKeyFields(new Field("key1", new IntType()), new Field("key2", new IntType()))
-                    .build());
-            StateStore stateStore = initialiseStateStore(new PartitionsBuilder(tableProperties)
-                    .singlePartition("A")
-                    .buildList());
-            IntStream.range(0, 10)
-                    .forEach(i -> ingestRecordsToSketchOnPartition(stateStore, "A",
-                            IntStream.range(0, 100)
-                                    // The majority of the values are 10; so min should equal median
-                                    .mapToObj(r -> new Record(Map.of(
-                                            "key1", r < 75 ? 10 : 20,
-                                            "key2", r)))));
-
-            // When
-            splitSinglePartition(stateStore, generateIds("B", "C"));
-
-            // Then
-            assertThat(stateStore.getAllPartitions())
-                    .containsExactlyInAnyOrderElementsOf(new PartitionsBuilder(tableProperties)
-                            .rootFirst("A")
-                            .splitToNewChildrenOnDimension("A", "B", "C", 1, 50)
-                            .buildList());
-        }
-
-        @Test
-        public void shouldSplitByteKeyOnFirstDimension() throws Exception {
-            // Given
-            tableProperties.setSchema(Schema.builder()
-                    .rowKeyFields(new Field("key1", new ByteArrayType()), new Field("key2", new ByteArrayType()))
-                    .build());
-            StateStore stateStore = initialiseStateStore(new PartitionsBuilder(tableProperties)
-                    .singlePartition("A")
-                    .buildList());
-
-            IntStream.range(0, 10)
-                    .forEach(i -> ingestRecordsToSketchOnPartition(stateStore, "A",
-                            IntStream.range(0, 100)
-                                    .mapToObj(r -> new Record(Map.of(
-                                            "key1", new byte[]{(byte) r},
-                                            "key2", new byte[]{(byte) -100})))));
-
-            // When
-            splitSinglePartition(stateStore, generateIds("B", "C"));
-
-            // Then
-            assertThat(stateStore.getAllPartitions())
-                    .containsExactlyInAnyOrderElementsOf(new PartitionsBuilder(tableProperties)
-                            .rootFirst("A")
-                            .splitToNewChildrenOnDimension("A", "B", "C", 0, new byte[]{50})
-                            .buildList());
-        }
-
-        @Test
-        public void shouldSplitByteKeyOnSecondDimensionWhenAllValuesForFirstKeyAreTheSame() throws Exception {
-            // Given
-            tableProperties.setSchema(Schema.builder()
-                    .rowKeyFields(new Field("key1", new ByteArrayType()), new Field("key2", new ByteArrayType()))
-                    .build());
-            StateStore stateStore = initialiseStateStore(new PartitionsBuilder(tableProperties)
-                    .singlePartition("A")
-                    .buildList());
-            IntStream.range(0, 10)
-                    .forEach(i -> ingestRecordsToSketchOnPartition(stateStore, "A",
-                            IntStream.range(0, 100)
-                                    .mapToObj(r -> new Record(Map.of(
-                                            "key1", new byte[]{(byte) -100},
-                                            "key2", new byte[]{(byte) r})))));
-
-            // When
-            splitSinglePartition(stateStore, generateIds("B", "C"));
-
-            // Then
-            assertThat(stateStore.getAllPartitions())
-                    .containsExactlyInAnyOrderElementsOf(new PartitionsBuilder(tableProperties)
-                            .rootFirst("A")
-                            .splitToNewChildrenOnDimension("A", "B", "C", 1, new byte[]{50})
-                            .buildList());
-        }
-    }
-
-    @Nested
-    @DisplayName("Commit partition split asynchronously")
-    class AsynchronousCommit {
-
-        @Test
-        void shouldCommitPartitionSplitAsynchronously() throws Exception {
-            // Given
-            PartitionTree tree = new PartitionsBuilder(tableProperties)
-                    .singlePartition("A")
-                    .buildTree();
-            StateStore stateStore = initialiseStateStore(tree.getAllPartitions());
-            String filename = ingestRecordsToSketchOnPartition(stateStore, "A",
-                    IntStream.rangeClosed(1, 100)
-                            .mapToObj(i -> new Record(Map.of("key", i))));
-            tableProperties.set(PARTITION_SPLIT_ASYNC_COMMIT, "true");
-            tableProperties.set(TABLE_ID, "tableId");
-
-            // When
-            partitionSplitter(stateStore, generateIds("B", "C"))
-                    .splitPartition(tree.getRootPartition(), List.of(filename));
-
-            // Then
-            assertThat(stateStore.getAllPartitions())
-                    .containsExactlyInAnyOrderElementsOf(tree.getAllPartitions());
-
-            PartitionTree resultant = new PartitionsBuilder(tableProperties)
-                    .rootFirst("A")
-                    .splitToNewChildren("A", "B", "C", 51)
-                    .buildTree();
-
-            assertThat(sentAsyncCommits).containsExactly(StateStoreCommitRequest.create("tableId",
-                    new SplitPartitionTransaction(resultant.getPartition("A"),
-                            List.of(resultant.getPartition("B"), resultant.getPartition("C")))));
-        }
-    }
-
-    private StateStore initialiseStateStore(List<Partition> partitions) {
-        return InMemoryTransactionLogStateStore.createAndInitialiseWithPartitions(partitions, tableProperties, new InMemoryTransactionLogs());
-    }
-
-    private String ingestRecordsToSketchOnPartition(StateStore stateStore, String partitionId, Stream<Record> recordsStream) {
-        Sketches sketches = Sketches.from(tableProperties.getSchema());
-        AtomicLong recordCount = new AtomicLong();
-
-        recordsStream.forEach(rec -> {
-            sketches.update(rec);
-            recordCount.incrementAndGet();
-        });
-
-        FileReference recordFileReference = FileReferenceFactory.from(instanceProperties, tableProperties, stateStore)
-                .partitionFile(partitionId, UUID.randomUUID().toString(), recordCount.get());
-        try {
-            update(stateStore).addFile(recordFileReference);
-        } catch (Exception ex) {
-            throw new RuntimeException(ex);
-        }
-
-        filenameToSketches.saveFileSketches(recordFileReference.getFilename(), tableProperties.getSchema(), sketches);
-        return recordFileReference.getFilename();
-    }
-
-    private void splitSinglePartition(StateStore stateStore, Supplier<String> generateIds) throws Exception {
-        Partition partition = stateStore.getAllPartitions().get(0);
-        List<String> fileNames = stateStore.getFileReferences().stream()
-                .map(FileReference::getFilename)
-                .collect(Collectors.toList());
-        SplitPartition partitionSplitter = partitionSplitter(stateStore, generateIds);
-        partitionSplitter.splitPartition(partition, fileNames);
-    }
-
-    private void splitPartition(StateStore stateStore, String partitionId, Supplier<String> generateIds) throws Exception {
-        PartitionTree tree = new PartitionTree(stateStore.getAllPartitions());
-        Partition partition = tree.getPartition(partitionId);
-        List<String> fileNames = stateStore.getFileReferences().stream()
-                .filter(file -> partitionId.equals(file.getPartitionId()))
-                .map(FileReference::getFilename)
-                .collect(Collectors.toList());
-        SplitPartition partitionSplitter = partitionSplitter(stateStore, generateIds);
-        partitionSplitter.splitPartition(partition, fileNames);
-    }
-
-    private SplitPartition partitionSplitter(StateStore stateStore, Supplier<String> generateIds) {
-        return new SplitPartition(stateStore, tableProperties, filenameToSketches, generateIds, sentAsyncCommits::add);
-    }
-
-    private static Supplier<String> generateIds(String... ids) {
-        return Arrays.stream(ids).iterator()::next;
-    }
-
-    private static Supplier<String> generateIdsStartingFrom(char start) {
-        return Stream.iterate(start, c -> (char) (c + 1)).map(String::valueOf).iterator()::next;
-    }
-
-    private static Supplier<String> generateNoIds() {
-        return () -> {
-            throw new IllegalArgumentException("Generated an ID, expected none");
-        };
-    }
-}
