+++ java/system-test/system-test-drivers/src/main/java/sleeper/systemtest/drivers/partitioning/AwsPartitionReportDriver.java
@@ -17,7 +17,7 @@
 package sleeper.systemtest.drivers.partitioning;
 
 import sleeper.clients.report.partitions.PartitionsStatusReporter;
-import sleeper.splitterv2.core.status.PartitionsStatus;
+import sleeper.splitter.core.status.PartitionsStatus;
 import sleeper.systemtest.dsl.instance.SystemTestInstanceContext;
 import sleeper.systemtest.dsl.reporting.PartitionReportDriver;
 import sleeper.systemtest.dsl.reporting.SystemTestReport;
