+++ java/system-test/system-test-dsl/pom.xml
@@ -66,7 +66,7 @@
         </dependency>
         <dependency>
             <groupId>sleeper</groupId>
-            <artifactId>splitterv2</artifactId>
+            <artifactId>splitter</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
         <!-- Test Dependencies -->
