+++ java/system-test/system-test-dsl/src/test/java/sleeper/systemtest/dsl/testutil/drivers/InMemoryPartitionSplittingDriver.java
@@ -21,9 +21,9 @@ import org.slf4j.LoggerFactory;
 import sleeper.core.properties.table.TableProperties;
 import sleeper.core.statestore.StateStore;
 import sleeper.sketchesv2.testutils.InMemorySketchesStore;
-import sleeper.splitterv2.core.find.FindPartitionsToSplit;
-import sleeper.splitterv2.core.find.FindPartitionsToSplit.JobSender;
-import sleeper.splitterv2.core.split.SplitPartition;
+import sleeper.splitter.core.find.FindPartitionsToSplit;
+import sleeper.splitter.core.find.FindPartitionsToSplit.JobSender;
+import sleeper.splitter.core.split.SplitPartition;
 import sleeper.systemtest.dsl.instance.SystemTestInstanceContext;
 import sleeper.systemtest.dsl.partitioning.PartitionSplittingDriver;
